src/main/java/org/example/Square.java
package org.example;

import java.util.Objects;

public class Square implements Shape {

    private double side;

    public Square(double side) {
        this.side = side;
    }

    public void setSide(double side) {
        this.side = side;
    }

    @Override
    public double calculatePerimeter() {
        return side * 4;
    }

    @Override
    public double calculateSurface() {
        return side * side;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj) {
            return true;
        }
        if (obj.getClass() != this.getClass()) {
            return false;
        }
        Square square = (Square) obj;

        return this.side == square.side;
    }

    @Override
    public int hashCode() {
        return Objects.hash(side);
    }
}
